Placed the type descriptions into a separate file
