+++ datatypes.py
@@ -1,0 +1,27 @@
+from dataclasses import dataclass
+from datetime import datetime
+from enum import Enum
+
+Celsius = float
+
+class WeatherType(str, Enum):
+    THUNDERSTORM = "Гроза"
+    DRIZZLE = "Изморось"
+    RAIN = "Дождь"
+    SNOW = "Снег"
+    CLEAR = "Ясно"
+    FOG = "Туман"
+    CLOUDS = "Облачно"
+
+@dataclass(slots=True, frozen=True)
+class Weather:
+    temperature: Celsius
+    weather_type: WeatherType
+    sunrise: datetime
+    sunset: datetime
+    city: str
+
+@dataclass(slots=True, frozen=True)
+class Coordinates:
+    latitude: float
+    longitude: float
